% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.rdpower}
\alias{summary.rdpower}
\alias{print.rdpower}
\alias{print.summary.rdpower}
\alias{summary.rdsampsi}
\alias{print.rdsampsi}
\alias{print.summary.rdsampsi}
\alias{summary.rdmde}
\alias{print.rdmde}
\alias{print.summary.rdmde}
\title{Summarize RDPOWER objects}
\usage{
\method{summary}{rdpower}(object, ..., all = FALSE)

\method{print}{rdpower}(x, ...)

\method{print}{summary.rdpower}(x, ...)

\method{summary}{rdsampsi}(object, ..., all = FALSE)

\method{print}{rdsampsi}(x, ...)

\method{print}{summary.rdsampsi}(x, ...)

\method{summary}{rdmde}(object, ..., all = FALSE)

\method{print}{rdmde}(x, ...)

\method{print}{summary.rdmde}(x, ...)
}
\arguments{
\item{object}{object returned by \code{\link{rdpower}()},
\code{\link{rdsampsi}()}, or \code{\link{rdmde}()}.}

\item{...}{additional arguments.}

\item{all}{displays the conventional output in addition to the robust
bias-corrected output.}

\item{x}{object returned by \code{\link{rdpower}()},
\code{\link{rdsampsi}()}, or \code{\link{rdmde}()}.}
}
\value{
A summary object containing the original return values, returned invisibly
by the print method.
}
\description{
\code{summary()} prints the default robust bias-corrected output. Use
\code{all = TRUE} to also display the conventional output.
}
